breast-cancer-backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
scikit-learn==1.6.1
numpy==2.2.3
pandas==2.2.3
gunicorn==23.0.0
python-dotenv==1.0.1
